backend/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.0
boto3==1.35.0
Pillow==10.4.0
sentry-sdk[fastapi]==2.14.0
slowapi==0.1.9
redis[hiredis]==5.1.0
openpyxl==3.1.5
reportlab==4.2.2
